+++ relecov_dashboard/views.py
@@ -8,17 +8,17 @@ import relecov_core.utils.handling_lineage
 import relecov_core.utils.handling_variant
 import relecov_dashboard.dashboard_config
 import relecov_dashboard.utils.generic.graphics.plotly
-import relecov_dashboard.utils.methodology.graphics.bioinfo
-import relecov_dashboard.utils.methodology.graphics.host_info
-import relecov_dashboard.utils.methodology.graphics.index
-import relecov_dashboard.utils.methodology.graphics.sample_preprocessing
-import relecov_dashboard.utils.methodology.graphics.sequencing
-import relecov_dashboard.utils.variants.graphics.heatmap_mutation_graph_by_lineage
-import relecov_dashboard.utils.variants.graphics.lineage_variation_over_time_graph
-import relecov_dashboard.utils.variants.graphics.lineages_in_time
-import relecov_dashboard.utils.variants.graphics.molecule3D_bn_graph
-import relecov_dashboard.utils.variants.graphics.needle_mutation_graph_by_lineage
-import relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie
+import relecov_dashboard.utils.methodology.graphics.met_bioinfo
+import relecov_dashboard.utils.methodology.graphics.met_host_info
+import relecov_dashboard.utils.methodology.graphics.met_index
+import relecov_dashboard.utils.methodology.graphics.met_sample_preprocessing
+import relecov_dashboard.utils.methodology.graphics.met_sequencing
+import relecov_dashboard.utils.variants.graphics.var_heatmap_mutation_graph_by_lineage
+import relecov_dashboard.utils.variants.graphics.var_lineage_variation_over_time_graph
+import relecov_dashboard.utils.variants.graphics.var_lineages_in_time
+import relecov_dashboard.utils.variants.graphics.var_molecule3D_bn_graph
+import relecov_dashboard.utils.variants.graphics.var_needle_mutation_graph_by_lineage
+import relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie
 
 
 # dashboard/variants
@@ -33,7 +33,7 @@ def mutations_in_lineage(request):
     def_chrom = relecov_core.utils.handling_variant.get_default_chromosome()
     lineages_list = relecov_core.utils.handling_lineage.get_lineages_list()
     mdata, lineage = (
-        relecov_dashboard.utils.variants.graphics.needle_mutation_graph_by_lineage.get_variant_data_from_lineages(
+        relecov_dashboard.utils.variants.graphics.var_needle_mutation_graph_by_lineage.get_variant_data_from_lineages(
             graphic_name="variations_per_lineage", lineage=None, chromosome=def_chrom
         )
     )
@@ -46,7 +46,7 @@ def mutations_in_lineage(request):
                 "ERROR": relecov_dashboard.dashboard_config.ERROR_NO_LINEAGES_ARE_DEFINED_YET
             },
         )
-    relecov_dashboard.utils.variants.graphics.needle_mutation_graph_by_lineage.create_needle_plot_graph_mutation_by_lineage(
+    relecov_dashboard.utils.variants.graphics.var_needle_mutation_graph_by_lineage.create_needle_plot_graph_mutation_by_lineage(
         lineages_list, lineage, mdata
     )
     return render(
@@ -59,7 +59,7 @@ def mutations_in_lineage(request):
 # FIXME: Couldn't find in variants dashboard a button or ref to acces this url
 @login_required
 def spike_mutations_3d(request):
-    relecov_dashboard.utils.variants.graphics.molecule3D_bn_graph.create_model3D_bn()
+    relecov_dashboard.utils.variants.graphics.var_molecule3D_bn_graph.create_model3D_bn()
     return render(request, "relecov_dashboard/variantSpikeMutations3D.html")
 
 
@@ -67,7 +67,7 @@ def lineages_voc(request):
     # Draw lineage based on time
     draw_lineages = {}
     draw_lineages["lineage_on_time"] = (
-        relecov_dashboard.utils.variants.graphics.lineage_variation_over_time_graph.create_lineages_variations_graphic()
+        relecov_dashboard.utils.variants.graphics.var_lineage_variation_over_time_graph.create_lineages_variations_graphic()
     )
     return render(
         request,
@@ -80,9 +80,9 @@ def lineages_voc(request):
 @login_required
 def samples_received_over_time_graph(request):
     df = (
-        relecov_dashboard.utils.variants.graphics.lineages_in_time.create_dataframe_from_json()
+        relecov_dashboard.utils.variants.graphics.var_lineages_in_time.create_dataframe_from_json()
     )
-    relecov_dashboard.utils.variants.graphics.lineages_in_time.create_samples_over_time_graph(
+    relecov_dashboard.utils.variants.graphics.var_lineages_in_time.create_samples_over_time_graph(
         df
     )
 
@@ -93,12 +93,12 @@ def samples_received_over_time_graph(request):
 @login_required
 def samples_received_over_time_pie(request):
     data = (
-        relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.parse_json_file()
+        relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.parse_json_file()
     )
-    relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.create_samples_received_over_time_per_ccaa_pieChart(
+    relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.create_samples_received_over_time_per_ccaa_pieChart(
         data
     )
-    relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.create_samples_received_over_time_per_laboratory_pieChart(
+    relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.create_samples_received_over_time_per_laboratory_pieChart(
         data
     )
 
@@ -109,12 +109,12 @@ def samples_received_over_time_pie(request):
 @login_required
 def samples_received_over_time_pie_laboratory(request):
     data = (
-        relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.parse_json_file()
+        relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.parse_json_file()
     )
-    relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.create_samples_received_over_time_per_ccaa_pieChart(
+    relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.create_samples_received_over_time_per_ccaa_pieChart(
         data
     )
-    relecov_dashboard.utils.variants.graphics.samples_received_over_time_pie.create_samples_received_over_time_per_laboratory_pieChart(
+    relecov_dashboard.utils.variants.graphics.var_samples_received_over_time_pie.create_samples_received_over_time_per_laboratory_pieChart(
         data
     )
 
@@ -165,7 +165,7 @@ def variants_mutations_in_lineages_heatmap(request):
             "relecov_dashboard/variantsMutationsInLineagesHeatmap.html",
             {"ERROR": relecov_core.core_config.ERROR_VARIANT_IN_SAMPLE_NOT_DEFINED},
         )
-    relecov_dashboard.utils.variants.graphics.heatmap_mutation_graph_by_lineage.create_heatmap(
+    relecov_dashboard.utils.variants.graphics.var_heatmap_mutation_graph_by_lineage.create_heatmap(
         sample_list, gene_list
     )
     return render(request, "relecov_dashboard/variantsMutationsInLineagesHeatmap.html")
@@ -174,7 +174,7 @@ def variants_mutations_in_lineages_heatmap(request):
 # dashboard/methodology
 @login_required
 def methodology_index(request):
-    graphics = relecov_dashboard.utils.methodology.graphics.index.index_dash_fields()
+    graphics = relecov_dashboard.utils.methodology.graphics.met_index.index_dash_fields()
     return render(
         request, "relecov_dashboard/methodologyIndex.html", {"graphics": graphics}
     )
@@ -183,7 +183,7 @@ def methodology_index(request):
 @login_required
 def methodology_host_info(request):
     host_info = (
-        relecov_dashboard.utils.methodology.graphics.host_info.host_info_graphics()
+        relecov_dashboard.utils.methodology.graphics.met_host_info.host_info_graphics()
     )
     if "ERROR" in host_info:
         return render(
@@ -197,7 +197,7 @@ def methodology_host_info(request):
 @login_required
 def methodology_sequencing(request):
     sequencing = (
-        relecov_dashboard.utils.methodology.graphics.sequencing.sequencing_graphics()
+        relecov_dashboard.utils.methodology.graphics.met_sequencing.sequencing_graphics()
     )
     if "ERROR" in sequencing:
         return render(
@@ -215,7 +215,7 @@ def methodology_sequencing(request):
 @login_required
 def methodology_sample_processing(request):
     sample_processing = (
-        relecov_dashboard.utils.methodology.graphics.sample_preprocessing.sample_processing_graphics()
+        relecov_dashboard.utils.methodology.graphics.met_sample_preprocessing.sample_processing_graphics()
     )
     if "ERROR" in sample_processing:
         return render(
@@ -232,7 +232,7 @@ def methodology_sample_processing(request):
 
 @login_required
 def methodology_bioinfo(request):
-    bioinfo = relecov_dashboard.utils.methodology.graphics.bioinfo.bioinfo_graphics()
+    bioinfo = relecov_dashboard.utils.methodology.graphics.met_bioinfo.bioinfo_graphics()
     return render(
         request, "relecov_dashboard/methodologyBioinfo.html", {"bioinfo": bioinfo}
     )
